refactor: Change directory name to camel type.
