+++ src/admin/pages/ContentType/Edit/EditField/fieldTypes/Boolean/index.tsx
